refactor: solve modal circluar dependency issue
